Move components to appropriate paths
